Add util module for version and rest
