+++ src/azure-cli/azure/cli/command_modules/util/commands.py
@@ -1,0 +1,13 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+
+def load_command_table(self, _):
+
+    with self.command_group('') as g:
+        g.custom_command('rest', 'rest_call')
+
+    with self.command_group('') as g:
+        g.custom_command('version', 'show_version')
