+++ src/azure-cli/azure/cli/command_modules/util/custom.py
@@ -1,0 +1,31 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+from __future__ import print_function
+
+from knack.log import get_logger
+
+logger = get_logger(__name__)
+
+
+def rest_call(cmd, uri, method=None, headers=None, uri_parameters=None,
+              body=None, skip_authorization_header=False, resource=None, output_file=None):
+    from azure.cli.core.util import send_raw_request
+    r = send_raw_request(cmd.cli_ctx, method, uri, headers, uri_parameters, body,
+                         skip_authorization_header, resource, output_file)
+    if not output_file and r.content:
+        try:
+            return r.json()
+        except ValueError:
+            logger.warning('Not a json response, outputting to stdout. For binary data '
+                           'suggest use "--output-file" to write to a file')
+            print(r.text)
+    return None
+
+
+def show_version(cmd):  # pylint: disable=unused-argument
+    from azure.cli.core.util import get_az_version_json
+    versions = get_az_version_json()
+    return versions
