+++ src/azure-cli/azure/cli/command_modules/util/tests/latest/recordings/test_rest.yaml
