+++ src/azure-cli/azure/cli/command_modules/util/tests/latest/rest_storage_account_create_body.json
